Ingresar ficha

+++ JavaLaberinto/src/Data/Ficha.java
@@ -7,13 +7,13 @@ public class Ficha {
     private char caracter; // me parecio mas adecuado nombrarlo caracter que tesoro por que no solo puede ser un tesoro
     private char[][] ficha;
     private ImageIcon imagen;
-    private int giro;
+    private double giro;
 
-    public int getGiro() {
+    public double getGiro() {
         return giro;
     }
 
-    public void setGiro(int giro) {
+    public void setGiro(double giro) {
         this.giro = giro;
     }
